+++ CHANGELOG.md
@@ -6,7 +6,8 @@
     - GetBucketV2 (ListObjectsV2): `Aliyun.Oss.Bucket.list_objects`
   - **Breaking changes**:
     - Use separated modules to scope SubResources of Bucket/Object, for example:
-      - `Aliyun.Oss.Bucket.get_lifecycle` becomes `Aliyun.Oss.Bucket.Lifecycle.get`
+      - `Aliyun.Oss.Bucket.get_acl` becomes `Aliyun.Oss.Bucket.ACL.get`
+      - `Aliyun.Oss.Object.get_acl` becomes `Aliyun.Oss.Object.ACL.get`
 
 ## v0.6.3
   - Use Supervised Task